refactor: create settings folder in shared to store components concern settingsWalletButton, remove unused import, etc #1
